rename: 파일명 변경 (postEnrollment -> Enrollment)
